fix: PR remarks, app rewrite
