+++ config.ru
@@ -2,8 +2,12 @@
 
 require_relative 'middleware/runtime'
 require_relative 'middleware/logger'
-require_relative 'app/application'
+require_relative 'apps/time_app'
+
+ROUTES = {
+  '/time' => TimeApp.new
+}.freeze
 
 use Runtime
 use AppLogger, logdev: File.expand_path('logs/app.log', __dir__)
-run Application.new
+run Rack::URLMap.new(ROUTES)
